requirements.txt
aiosmtplib==3.0.2 ; python_version >= "3.12" and python_version < "4.0"
aiosqlite==0.21.0 ; python_version >= "3.12" and python_version < "4.0"
alembic==1.14.1 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.8.0 ; python_version >= "3.12" and python_version < "4.0"
asyncpg==0.30.0 ; python_version >= "3.12" and python_version < "4.0"
bcrypt==3.2.2 ; python_version >= "3.12" and python_version < "4.0"
blinker==1.9.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.12" and python_version < "4.0"
cloudinary==1.42.2 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows" or python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
coverage==7.6.10 ; python_version >= "3.12" and python_version < "4.0"
deprecated==1.2.18 ; python_version >= "3.12" and python_version < "4.0"
dnspython==2.7.0 ; python_version >= "3.12" and python_version < "4.0"
ecdsa==0.19.0 ; python_version >= "3.12" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
fastapi-mail==1.4.2 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.115.8 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.1.1 ; python_version >= "3.12" and python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.5 ; python_version >= "3.12" and python_version < "4.0"
libgravatar==1.0.4 ; python_version >= "3.12" and python_version < "4.0"
limits==4.0.1 ; python_version >= "3.12" and python_version < "4.0"
mako==1.3.9 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.12" and python_version < "4.0"
passlib==1.7.4 ; python_version >= "3.12" and python_version < "4.0"
redis==5.0.1 ; python_version >= "3.12" and python_version < "4.0"hon_version < "4.0"
pluggy==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
psycopg2==2.9.10 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.27.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic-settings==2.7.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.10.6 ; python_version >= "3.12" and python_version < "4.0"
pytest-asyncio==0.25.3 ; python_version >= "3.12" and python_version < "4.0"
pytest-cov==6.0.0 ; python_version >= "3.12" and python_version < "4.0"
pytest==8.3.4 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
python-jose==3.3.0 ; python_version >= "3.12" and python_version < "4.0"
python-multipart==0.0.6 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
slowapi==0.1.9 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.38 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.45.3 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.3.0 ; python_version >= "3.12" and python_version < "4.0"
uvicorn==0.34.0 ; python_version >= "3.12" and python_version < "4.0"
wrapt==1.17.2 ; python_version >= "3.12" and python_version < "4.0"
